// Segundo Parcial/Trabajo complementario-3/crudTutoria/src/interfaces/ITutoria.ts
// Generated by https://quicktype.io

export interface IResTutoria {
    sum:     number;
    tutorias: Tutoria[];
}

export interface Tutoria {
    _id?:    string;
    asignatura: string;
    status?: boolean;
    numhoras: string;
    fecha:  string;
    hora:  string;
    category?:  number;
    __v?:    number;
}